+++ src/main.ts
@@ -2,6 +2,8 @@ import Phaser from 'phaser';
 import BootScene from './scenes/BootScene';
 import GameScene from './scenes/GameScene';
 import Level2Scene from './scenes/Level2Scene';
+import Level2HardScene from './scenes/Level2HardScene';
+import TransitionScene from './scenes/TransitionScene';
 import TestScene from './scenes/TestScene';
 
 import StartScene from './scenes/StartScene';
@@ -20,10 +22,10 @@ const config: Phaser.Types.Core.GameConfig = {
         default: 'arcade',
         arcade: {
             gravity: { x: 0, y: 0 },
-            debug: true
+            debug: false
         }
     },
-    scene: [BootScene, StartScene, GameScene, Level2Scene, TestScene]
+    scene: [BootScene, StartScene, GameScene, TransitionScene, Level2HardScene, Level2Scene, TestScene]
 };
 
 new Phaser.Game(config);